From education: removed fields: attendedFrom, attendedTo Merged applyingFor into account page, merged stadardTest page to educations page Not going to use the Checklist entity and dependents for now. Using evaluation class to generate checklist page. created checklist report, by splitting the evaluations report.
After an admissionOfficer created the account, 
redirect user to the profile page of the newly registered student
		 

When AO creates an account for an applicant.
Auto generate username for the applicant.
Format the username as: firstName + a 3-digit number.
Make sure the username is not already taken.

Using JDBC to update userEntity
If use hibernate persistence, the form is not validates as the
agreeTerms must be 'true'
Only, student can make the agreement. AO can't fill in this field.
When AO creates an account for an applicant.
Auto generate unique username for the applicant.
Make sure the email is not already taken too.

added new role: ROLE_STUDENT_DENIED