Using JDBC to update userEntity When AO creates an account for an applicant. Auto generate unique username for the applicant. Make sure the email is not already taken too. If use hibernate persistence, the form is not validates as the agreeTerms must be 'true' Only, student can make the agreement. AO can't fill in this field.
If email is available then persist user
When Admission Officer creates a user account: acceptTerms should not be
'true'. 
ONLY applicant can agree to terms by him/herself
		 

Admission Officer to create a student's account
When you load the registration form. set the acceptTerms "true".
If AO regsiters a student, the student must have signed a form 
stating the applicant has read the privacy policy and that the applicant
agrees with it.


Has to do with acceptTerms??
editAccount

		/**
		 * Check if user has already agreed to the terms.
		 * get acceptTerms value form the DB.
		 * If user has accepted terms, leave it as it is.
		 * 
		 * Else,
		 * If account was created by AO, and not yet updated by the student. 
		 * then the acceptTerms is stil in false.
		 * To pass validation acceptTerms is set to true in the GET method.
		 * Now, revert it back to False.
		 * 
		 */