created new page: privacyAndPolicyStatement
When an AO create a student's account
in getRegisterUserForm()

When you load the registration form. set the acceptTerms "true".
If AO registers an applicant, the applicant must have signed a form 
stating the applicant has read the privacy policy and that the applicant
agrees with it. So, replaced method insertUserEntity(form, result) by
method createUserEntityAsAO(form, result) to create an account.