added null value options in the form:select tag Allow applicant to make changes - this button in accounts page allows Admission Officers to change the state of the applicant after it is closed so that applicant can make changes to application.
Country - the first item in country, i.e. ID=1 is the null country. this
represents when a user does’t select country. till i fix the null value
in option.

When using
<form:select path="citizenship.id">
<form:option value="" label="" />

The page doesn’t parse the null value option. a null value validation is
not caught by the result binding result;